FINERACT-1971: Fix for not properly resolving delinquency range data for loan in case there's an installment delinquency
